Cleanup excess styles
